wigner: 6j integer sum check was failing
In addition, I added some code to detail which triads were failing the test.
The text is added to the `msg` string with a single call to print if any of the
tests fail.

It would be nice to have some global verbosity flag to allow the user to turn
on/off the message.

Finally, I offer two error messages (coincidentally) of the same length.